Add a simple mailing list example.
This works, but does not handle any persistence, and is spitting back raw body
content which doesn't work correctly with HTML emails. Will look into how do do
this in Mail. This code is telling us that the manual filtering of the email
sender names is tedious too, will need to work ideas out around that soon. Also,
we should set the sender so that it's the original sender, not the mailing list
address.