Build a simple clone of libre list's automatic mailing list generation functionality.
This adds support for named matches in conditions, but also introduces a lot of
ugliness and complexity that I want to clean up. Except for adding a few bolt-on
features, I don't think I should add more core bits of functionality before
adding some proper tests and doing a bit of refactoring.