Major overhaul of the configuration system.
It's important to note this is a *starting point* only. I plan to add some
basic guards against typos, and some clever support for mail provider profiles.
With the exception of perhaps the application specific configuration, the use of
ostruct will pretty much go away. Still, it gives us the API we want for now.

I wonder how the Ruby community is going to take Lua style configuration files.